% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_traj_identifier}
\alias{add_traj_identifier}
\title{Add a unique identifier to each trajectory}
\usage{
add_traj_identifier(
  trajdata,
  ntraj_1 = FALSE,
  direction = "forward",
  openair_compliant = TRUE
)
}
\arguments{
\item{trajdata}{a trajectory dataframe returned from splitr::trajectory_read.}

\item{ntraj_1}{TRUE/FALSE is there only one trajectory per day?}

\item{direction}{hysplit trajectory model run direction. One of either 'forward' or 'backward'.}

\item{openair_compliant}{TRUE/FALSE Has the trajectory dataframe already been converted to openair format with convert_openair? Set to FALSE if not.}
}
\description{
Add a unique numerical identifier to each trajectory for straightforward parsing
}
