% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfunctions.R
\name{archive_cluster_endpoints}
\alias{archive_cluster_endpoints}
\title{Archive all endpoints currently in the cluster endpoints directory}
\usage{
archive_cluster_endpoints(
  cluster_endpoints_dir = "C:/hysplit/cluster/endpts/",
  cluster_endpoints_archive_dir = "C:/hysplit/cluster/endpts/archive/",
  archive_name = "endpts",
  stopifempty = FALSE,
  append_timestamp = TRUE
)
}
\arguments{
\item{cluster_endpoints_dir}{character vecotor. Full file path to the cluster endpoints directory.}

\item{cluster_endpoints_archive_dir}{character vector. Full file path to the desired archive directory.}

\item{archive_name}{character vector to be added to the folder containing the archived endpoints.}

\item{stopifempty}{TRUE or FALSE to stop the operation if there are no endpoints to archive.}

\item{append_timestamp}{TRUE or FALSE to add a timestamp of the archive operation.}
}
\description{
Creates a folder within the specified archive directory, and moves all non-default files currently in the
cluster endpt directory to that folder.
}
