% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-analysis.R
\name{cluster_annual_freqs}
\alias{cluster_annual_freqs}
\title{Generate annual frequencies and stacked frequencies of clusters}
\usage{
cluster_annual_freqs(
  trajdata_frame,
  cluslist,
  stepfmt = FALSE,
  year_range = NULL,
  verbose = TRUE
)
}
\arguments{
\item{trajdata_frame}{A dataframe containing trajectories to be associated with those in the cluslist. A multi-trajectory output from trajectory_read() or a similar function. See example package data for an example data frame.}

\item{cluslist}{A cluslist - an output from \code{read_cluslist()}.}

\item{stepfmt}{TRUE/FALSE to format data for a geom_step() in ggplot. Will add an additional observation at the tail of each cluster's frequencies.}

\item{year_range}{an optional sequence of years (e.g., seq(2005,2020,1) or c(2005,2020)) to subset the trajdata.}

\item{verbose}{TRUE/FALSE to print messages as the function proceeds.}
}
\description{
Generate annual frequencies (\%) for each cluster within a cluster analysis,
along with stacked cluster data that allows for plotting as a single unit.
}
