% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convert_openair}
\alias{convert_openair}
\title{Convert a trajectory data frame into openair format}
\usage{
convert_openair(trajdata)
}
\arguments{
\item{trajdata}{a trajectory dataframe returned from splitr::trajectory_read}
}
\description{
Convert a standard trajectory dataframe from splitR to a format compatible with the openair package
}
