% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfunctions.R
\name{create_clusmeans_analogue}
\alias{create_clusmeans_analogue}
\title{Create an analogue cluster mean trajectory from one or more supplied cluslists.}
\usage{
create_clusmeans_analogue(
  clustlists,
  export_directory = NULL,
  endpoint_folder,
  traj_total_duration_hrs = 120
)
}
\arguments{
\item{clustlists}{one or more cluslist files, such as a direct output from read_cluslist() or get_cluslists().}

\item{export_directory}{NULL or full file path to a directory. If !NULL, mean trajectory tables will be exported as .csvs.}

\item{endpoint_folder}{full file path to a folder containing the endpoints used in the cluster calculations. If this is different to the file paths found in the cluslist, the endpoint folder will be used instead.}

\item{traj_total_duration_hrs}{numeric; how long were the trajectories run for, in hours?}
}
\description{
Experimental function. Takes cluslist files and a set of corresponding trajectories,
and attempts to create a set of clusmean files. The mean calculation uses geosphere::geomean, but
produces outputs that differ slightly from HYSPLIT's trajmean program. I could not diagnose
the differences, but they seem to be minor. Bears further investigation another time!
}
