% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{find_new_endpts}
\alias{find_new_endpts}
\title{Find endpoint files that do not match.}
\usage{
find_new_endpts(
  prev_endpt_filenames,
  new_endpt_filenames,
  exclude_date_bounds = TRUE,
  skip = "2 days"
)
}
\arguments{
\item{prev_endpt_filenames}{a vector of filenames to check against.}

\item{new_endpt_filenames}{a vector of filenames to identify new files within (those not in the prev_endpt_filenames).}

\item{exclude_date_bounds}{TRUE/FALSE to simply identify new trajectories based on whether they occur after the last date in the previous trajectories.}

\item{skip}{A character vector used to generate a date sequence, e.g. "2 days". This is equivalent to the skip field in the HYSPLIT cluster program. Note that there are some differences - for some years, the HYSPLIT cluster program seems to assume that September has 31 days, for instance.}
}
\description{
Experimental function. Take two trajectory endpoint filename vectors (old and new) and
find those that are not present in the former. Intended for use where new
endpoint files are added to a pre-existing folder containing e.g., endpoint
files used for a previous cluster analysis.
}
