% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfunctions.R
\name{format_obj_DELPCT}
\alias{format_obj_DELPCT}
\title{Format a DELPCT file for plotting}
\usage{
format_obj_DELPCT(DELPCT_table, threshold = 20)
}
\arguments{
\item{DELPCT_table}{an output from read_DELPCT()}

\item{threshold}{numeric value for \% change used to calculate whether an n cluster decrease exceeds a threshold value for reduced total spatial variance.}
}
\description{
Tore initial bits out of trajSpatial::plot_DELPCT. Useful for more refined line plotting.
}
