% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfunctions.R
\name{format_trajdata_clusterfreqs}
\alias{format_trajdata_clusterfreqs}
\title{Assign trajectories to their clusters within a trajectory data table.}
\usage{
format_trajdata_clusterfreqs(trajectory_data, cluslist, new_assignments = NULL)
}
\arguments{
\item{trajectory_data}{a data.frame of trajectory data, such as that obtained via splitr::trajectory_read().}

\item{cluslist}{a single cluslist file, such as a direct output from read_cluslist() or an individual cluslist from get_cluslists().}

\item{new_assignments}{NULL; Experimental; new numbers for recoding.}
}
\description{
Attach cluster numbers to trajectories within a trajectory data table, such as that obtained with splitr::trajectory_read().
}
