% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-analysis.R
\name{get_class_percentages}
\alias{get_class_percentages}
\title{Calculate percentage importance of clusters grouped into classes}
\usage{
get_class_percentages(cluslist, assignments)
}
\arguments{
\item{cluslist}{a single cluslist file, such as a direct output from read_cluslist() or an individual cluslist from get_cluslists().}

\item{assignments}{a data.frame containing the columns: 'new_clust' containing the cluster number from 1 to n clusters, 'class' containing the names of the classes, and 'class_num' containing numeric numbers for each class.}
}
\description{
Experimental function. A simple table of each cluster alongside columns containing 'class' and 'class_num' can be used
to group clusters into classes for grouped analysis. See Markle et al. (2012) for an example of this type of analysis.
}
