% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-parsers.R
\name{get_cluslists}
\alias{get_cluslists}
\title{Read a set of cluslist files}
\usage{
get_cluslists(cluster_dir = "C:/hysplit/cluster/working/")
}
\arguments{
\item{cluster_wd}{a full file path to the cluster working directory. Defaults to "C:/hysplit/cluster/working/".}
}
\description{
Read a set of cluslist files. Individual cluslist files are generated by the HYSPLIT clustering program when a clustering attempt is performed
for a given n cluster value, and contains a table listing each individual endpoint file and the cluster they have been assigned to.
}
