% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfunctions.R
\name{get_clusmeans_tdump}
\alias{get_clusmeans_tdump}
\title{Read a set of cluster tdump files}
\usage{
get_clusmeans_tdump(
  cluster_dir = "C:/hysplit/cluster/working/",
  type = "hysplit",
  header_string = "RESSURE"
)
}
\arguments{
\item{cluster_dir}{a full filepath to the cluster working directory. Defaults to "C:/hysplit/cluster/working/".}

\item{type}{experimental; one of either 'hysplit' or 'analogue' to specify either a hysplit program-generated file or a custom cluster generated in R using other functions in the trajSpatial package.}

\item{header_string}{character vector used to target the header line within the tdump file}
}
\description{
Read in all cluster tdump files (i.e. mean cluster trajectory information) in the specified directory. This can be used
to collate all mean cluster information after a set of clustering runs for analysis or display.
}
