% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfunctions.R
\name{get_cluster_percentages}
\alias{get_cluster_percentages}
\title{Calculate the percentage of endpoints in each cluster}
\usage{
get_cluster_percentages(cluslist)
}
\arguments{
\item{cluslist}{a single cluslist file, such as a direct output from read_cluslist() or an individual cluslist from get_cluslists().}
}
\description{
Using a supplied cluslist file, derives the percentage of endpoints contained in each cluster, providing a means
to determine their relative importance.
}
