% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-prep.R
\name{make_INFILE}
\alias{make_INFILE}
\title{Make an INFILE for a cluster analysis.}
\usage{
make_INFILE(
  endpoint_dir = "C:/hysplit/cluster/endpts/",
  cluster_wd = "C:/hysplit/cluster/working/",
  base_name = NULL,
  flag_longfilenames = FALSE,
  disallowed_flags = "readme",
  export = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{endpoint_dir}{A full path to the directory containing the endpoints to be used for clustering.}

\item{cluster_wd}{The cluster working directory. Defaults to the standard location for a windows default install.}

\item{base_name}{A string used to subset endpoint files. NULL by default, so all endpoints in the supplied directory are used.}

\item{flag_longfilenames}{TRUE/FALSE to print message if any endpoints have a filename exceeding 54 characters.}

\item{disallowed_flags}{character strings supplied as a character vector to be used to match files to be excluded from the INFILE.}

\item{export}{TRUE/FALSE to write the INFILE to the supplied cluster working directory.}

\item{verbose}{TRUE/FALSE to print status message}
}
\description{
The INFILE is created during the clustering setup and lists all the endpoints to be supplied to the clustering program.
}
