% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-parsers.R
\name{read_cluslist}
\alias{read_cluslist}
\title{Read a single cluslist file}
\usage{
read_cluslist(CLUSLIST_file)
}
\arguments{
\item{CLUSLIST_file}{a full file path to the desired cluslist file.}
}
\description{
Read a single cluslist file. The cluslist file is generated by the HYSPLIT clustering program when a clustering attempt is performed
for a given n cluster value, and contains a table listing each individual endpoint file and the cluster they have been assigned to. This function
is a simple wrapper for read.table().
}
