% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfunctions.R
\name{read_DELPCT}
\alias{read_DELPCT}
\title{Read in cluster change-in-TSV file.}
\usage{
read_DELPCT(
  cluster_wd = "C:/hysplit/working/cluster/",
  max_clusters = NULL,
  verbose = F
)
}
\arguments{
\item{cluster_wd}{full file path to the clustering working directory. It is set to 'C:/hysplit/cluster/working/' by default.}

\item{max_clusters}{NULL or numeric to trim the returned data frame to only contain less or equal cluster quantities than the number specified.}
}
\description{
A HYSPLIT clustering attempt will output a file containing data on the
percent change in total spatial variance or 'change in TSV \%'. This file is called
'DELPCT' and can be found in the cluster working directory after a given clustering attempt.
This function reads in this file, and optionally trims it to above a given n_clusters.
}
