% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-parsers.R
\name{read_endpoint_file}
\alias{read_endpoint_file}
\title{Read a single endpoint file}
\usage{
read_endpoint_file(file, header_string = "PRESSURE")
}
\arguments{
\item{file}{a full file path to the desired endpoint file.}

\item{header_string}{a character string used to target the header line in the file.}
}
\description{
Read and format a single endpoint (trajectory) file. An endpoint file is generated for every run of the hysplit model, and contains
information for individual trajectories or releases. This is a very simple alternative to \code{splitr::trajectory_read()}
}
