% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-parsers.R
\name{read_meanclus_tdump}
\alias{read_meanclus_tdump}
\title{Read a single cluster tdump mean trajectory file}
\usage{
read_meanclus_tdump(
  meanclus_tdump,
  header_string = "RESSURE",
  type = "hysplit"
)
}
\arguments{
\item{meanclus_tdump}{a full filepath to a cluster tdump file}

\item{header_string}{character vector used to target the header line within the tdump file}

\item{type}{experimental; one of either 'hysplit' or 'analogue' to specify either a hysplit program-generated file or a custom cluster generated in R using other functions in the trajSpatial package.}
}
\description{
Read a single cluster tdump file. The cluster tdump file is generated in the HYSPLIT clustering GUI when the
user presses "Display Means" for a given n clusters after clustering computations are performed. It contains information
that can be used to plot or analyse the mean cluster trajectories.
}
