% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfunctions.R
\name{recode_cluslist}
\alias{recode_cluslist}
\title{Recoding cluslist table number assignments}
\usage{
recode_cluslist(cluslist, old_clust, new_clust)
}
\arguments{
\item{cluslist}{a single cluslist file, such as a direct output from read_cluslist() or an individual cluslist from get_cluslists().}

\item{old_clust}{numeric vector of the old cluster numbers, in the order they appear.}

\item{new_clust}{numeric vector of the new desired cluster numbers, in the order they will replace the old_clust values.}
}
\description{
The numbering of clusters is determine by the algorithm, and as such can be a bit of a nightmare when plotting.
This function offers a simple way to recode the number of each cluster in a cluslist file.
}
