% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-analysis.R
\name{recode_clusmean}
\alias{recode_clusmean}
\title{Recoding clusmean table number assignments}
\usage{
recode_clusmean(clusmean, old_clust, new_clust)
}
\arguments{
\item{clusmean}{a single clusmean file, such as a direct output from read_meanclus_tdump().}

\item{old_clust}{numeric vector of the old cluster numbers, in the order they appear.}

\item{new_clust}{numeric vector of the new desired cluster numbers, in the order they will replace the old_clust values.}
}
\description{
The numbering of clusters is determined by the algorithm, and as such can be a bit of a nightmare when plotting.
This function offers a simple way to recode the number of each cluster in a clusmean file.
}
