% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-retro.R
\name{retro_cluster}
\alias{retro_cluster}
\title{Retrospectively assign new trajectories to existing clusters.}
\usage{
retro_cluster(
  trajectory_filenames,
  clusmean,
  time_interval_hrs = 12,
  verbose = TRUE
)
}
\arguments{
\item{trajectory_filenames}{a vector of filenames for trajectories.}

\item{clusmean}{a clusmean file. Can be either a tdump file from HYSPLIT or an 'analogue' generated by trajSpatial. Returned from \code{create_clusmeans_analogue()}.}

\item{time_interval_hrs}{the time interval used to select endpoints for comparison, where 1 = all endpoints and 12 = endpoints every 12 hours.}

\item{verbose}{TRUE/FALSE to show progress as calculation proceeds.}
}
\description{
Determine which cluster/s a set of trajectories are most similar to. Uses
geosphere::geoDist to determine the cluster with the least total distance between
trajectory and cluster endpoints at the specified time interval. A data frame
of the trajectories (filenames) and their corresponding cluster is returned.
Originally written to allow for clustering in cases where the size of a trajectory dataset
exceeds the ability of the HYSPLIT program to cluster it. In such a case, a representative
portion of a trajectory dataset (the vast majority, ideally) can be clustered and then the
rest can be retrospectively added.
}
