% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_trajdata_shp.R
\name{trim_trajdata_shp}
\alias{trim_trajdata_shp}
\title{Archive all endpoints currently in the cluster endpoints directory}
\usage{
trim_trajdata_shp(shapefile, trajdata, xy = c("lon", "lat"), verbose = TRUE)
}
\arguments{
\item{shapefile}{an ARCGIS shapefile (.shp) loaded into R with e.g. readOGR}

\item{trajdata}{a data frame of trajectory endpoints containing lat/lon data of each endpoint}

\item{xy}{character vector of the column names containing the x (lon) and y (lat) data within the trajdata data frame.}

\item{verbose}{TRUE/FALSE to display progress messages to help with error diagnostics.}
}
\description{
Trims a trajectory data frame to only those points lying within the
bounds of a shapefile polygon.
}
