% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_trajdata_shp_metrics.R
\name{trim_trajdata_shp_metrics}
\alias{trim_trajdata_shp_metrics}
\title{Calculate various metrics for trajectory data portions by trimming to a shapefile.}
\usage{
trim_trajdata_shp_metrics(
  shapefile,
  trajdata,
  xy = c("lon", "lat"),
  type = "month",
  verbose = TRUE
)
}
\arguments{
\item{shapefile}{an ARCGIS shapefile (.shp) loaded into R with e.g. readOGR}

\item{trajdata}{a data frame of trajectory endpoints containing lat/lon data of each endpoint}

\item{xy}{character vector of the column names containing the x (lon) and y (lat) data within the trajdata data frame.}

\item{type}{character; one of 'month', 'month_seq', or 'year'}

\item{verbose}{TRUE or FALSE to display progress messages to help with error diagnostics.}
}
\description{
Trims a trajectory data frame to only those points lying within the
bounds of a shapefile polygon, and then calculates the mean and standard deviation
of various endpoint parameters: height, pressure, lat, lon.
}
