% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-retro.R
\name{update_cluslist}
\alias{update_cluslist}
\title{Update a cluslist with new trajectories}
\usage{
update_cluslist(cluslist, retro_clustered)
}
\arguments{
\item{cluslist}{A cluslist object.}

\item{retro_clustered}{Result returned from \code{retro_cluster()}.}
}
\description{
Append trajectories to an existing cluslist.
}
