% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-parsers.R
\name{get_clusmeans_tdump}
\alias{get_clusmeans_tdump}
\title{Read a set of cluster tdump files}
\usage{
get_clusmeans_tdump(
  cluster_dir = "C:/hysplit/cluster/working/",
  type = "hysplit",
  header_string = "RESSURE"
)
}
\arguments{
\item{cluster_dir}{a full filepath to the cluster working directory. Defaults to "C:/hysplit/cluster/working/".}

\item{type}{One of either 'hysplit' or 'analogue' to specify either a hysplit program-generated file or a cluster mean tdump file generated in R using \code{create_clusmeans_analogue()}.}

\item{header_string}{character vector used to target the header line within the tdump file}
}
\description{
Read in all cluster tdump files (i.e. mean cluster trajectory information) in the specified directory, using an
iterative application of \code{read_meanclus_tdump()}. This can be used to collate all mean cluster information after a set of clustering runs for analysis or display.
}
